done with user creation and file deployed

+++ src/utils/cloundinary.js
@@ -1,9 +1,9 @@
 import { v2 as cloudinary } from 'cloudinary';
 import fs from "fs"//file system
     cloudinary.config({ 
-        cloud_name:process.env.cloud_name, 
-        api_key: process.env.cloud_api_key, 
-        api_secret: process.env.clound_api_secret // Click 'View API Keys' above to copy your API secret
+        cloud_name: process.env.cloud_name,
+        api_key: process.env.cloud_api_key,
+        api_secret: process.env.cloud_api_secret // Click 'View API Keys' above to copy your API secret
     });
     
 const Uploadon=async(localfilepath)=>{
@@ -12,20 +12,16 @@ const Uploadon=async(localfilepath)=>{
         const response=await cloudinary.uploader.upload(localfilepath,{
             resource_type:"auto"
         })
-        console.log("file is uploaded on cloudinary",response.url)
+        //console.log("file is uploaded on cloudinary",response.url)
+        fs.unlinkSync(localfilepath)
         return response
     }catch(error){
-        fs.unlinkSync(localfilepath)
+        console.log("CLOUDINARY ERROR:", error);
+        if (localfilepath && fs.existsSync(localfilepath)) {
+            fs.unlinkSync(localfilepath);
+        }
         // remove the locally saved temporary file as upload operation got failed
         return null;
     }
 }
 export { Uploadon }
-    cloudinary.uploader.upload(
-           'https://res.cloudinary.com/demo/image/upload/getting-started/shoes.jpg', {
-               public_id: 'shoes',
-           }
-       )
-       .catch((error) => {
-           console.log(error);
-       });
